Refactor decompiler to handle some control flow statements
This commit refactors the decompiler to handle control flow statements such as if, if-else, while loop and switch. The `Statement` enum has been updated to include these new types of statements. The `decompile_node` function has also been updated to take in a `ControlFlow` object which is used to determine the current node being processed.