Add support for struct types in the decompiler stack
This commit adds support for struct types to the decompiler stack. It also includes changes to the `StackEntry` enum and `BinaryOperator` enum, as well as updates to the C++ formatter. The new `Struct` variant of `StackEntry` represents a struct type with an origin and size, while the new `LogicalAnd` and `LogicalOr` variants of `BinaryOperator` represent logical AND and OR operations respectively. The C++ formatter has been updated to handle formatting of struct types in the stack.