Add type deduction
Added support for type deduction in the decompiler by introducing a new `LinkedValueType` struct that allows for linked lists of types. This change also includes updates to the `DecompiledFunction`, `Statement`, and `Function` structs to use this new type, as well as changes to the stack entry types to use it. Additionally, several binary and unary operators were updated to handle different types of values.